requirements.txt
# Core dependencies
openai>=1.0.0
anthropic>=0.8.0
pydantic>=2.0.0
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# API dependencies
aiohttp>=3.9.0
asyncio-throttle>=1.0.2

# Data handling
pydantic-settings>=2.0.0
jsonschema>=4.19.0

# CLI and UI
click>=8.1.0
rich>=13.6.0
typer>=0.9.0

# Logging and monitoring
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development
black>=23.9.0
flake8>=6.1.0
mypy>=1.6.0
